refactor: new divider component
