+++ packages/appkit/exports/constants.ts
@@ -1,1 +1,1 @@
-export const PACKAGE_VERSION = '1.4.1'
+export const PACKAGE_VERSION = '1.5.0'
