+++ packages/ui-new/src/components/wui-divider/index.ts
@@ -1,12 +1,12 @@
 import { html, LitElement } from 'lit'
 import { property } from 'lit/decorators.js'
-import '../../components/wui-text/index.js'
+import '../wui-text/index.js'
 import { resetStyles } from '../../utils/ThemeUtil.js'
 import { customElement } from '../../utils/WebComponentsUtil.js'
 import styles from './styles.js'
 
-@customElement('wui-separator')
-export class WuiSeparator extends LitElement {
+@customElement('wui-divider')
+export class WuiDivider extends LitElement {
   public static override styles = [resetStyles, styles]
 
   // -- State & Properties -------------------------------- //
@@ -20,7 +20,7 @@ export class WuiSeparator extends LitElement {
   // -- Private ------------------------------------------- //
   private template() {
     if (this.text) {
-      return html`<wui-text variant="small-500" color="fg-200">${this.text}</wui-text>`
+      return html`<wui-text variant="md-regular" color="inherit">${this.text}</wui-text>`
     }
 
     return null
@@ -29,6 +29,6 @@ export class WuiSeparator extends LitElement {
 
 declare global {
   interface HTMLElementTagNameMap {
-    'wui-separator': WuiSeparator
+    'wui-divider': WuiDivider
   }
 }
