--- packages/ui-new/src/layout/wui-separator/styles.ts
@@ -1,19 +1,0 @@
-import { css } from 'lit'
-
-export default css`
-  :host {
-    position: relative;
-    display: flex;
-    width: 100%;
-    height: 1px;
-    background-color: var(--wui-color-gray-glass-005);
-    justify-content: center;
-    align-items: center;
-  }
-
-  :host > wui-text {
-    position: absolute;
-    padding: 0px 10px;
-    background-color: var(--wui-color-modal-bg);
-  }
-`
